Workers/report-analyzer/requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
langchain_community